Move HydrawiseEntity into entity.py
