+++ homeassistant/components/hydrawise/__init__.py
@@ -6,10 +6,9 @@ import voluptuous as vol
 
 from homeassistant.components import persistent_notification
 from homeassistant.const import CONF_ACCESS_TOKEN, CONF_SCAN_INTERVAL
-from homeassistant.core import HomeAssistant, callback
+from homeassistant.core import HomeAssistant
 import homeassistant.helpers.config_validation as cv
-from homeassistant.helpers.dispatcher import async_dispatcher_connect, dispatcher_send
-from homeassistant.helpers.entity import Entity, EntityDescription
+from homeassistant.helpers.dispatcher import dispatcher_send
 from homeassistant.helpers.event import track_time_interval
 from homeassistant.helpers.typing import ConfigType
 
@@ -73,33 +72,3 @@ class HydrawiseHub:
     def __init__(self, data):
         """Initialize the entity."""
         self.data = data
-
-
-class HydrawiseEntity(Entity):
-    """Entity class for Hydrawise devices."""
-
-    _attr_attribution = "Data provided by hydrawise.com"
-
-    def __init__(self, data, description: EntityDescription) -> None:
-        """Initialize the Hydrawise entity."""
-        self.entity_description = description
-        self.data = data
-        self._attr_name = f"{self.data['name']} {description.name}"
-
-    async def async_added_to_hass(self):
-        """Register callbacks."""
-        self.async_on_remove(
-            async_dispatcher_connect(
-                self.hass, SIGNAL_UPDATE_HYDRAWISE, self._update_callback
-            )
-        )
-
-    @callback
-    def _update_callback(self):
-        """Call update method."""
-        self.async_schedule_update_ha_state(True)
-
-    @property
-    def extra_state_attributes(self):
-        """Return the state attributes."""
-        return {"identifier": self.data.get("relay")}
