+++ homeassistant/components/hydrawise/sensor.py
@@ -16,8 +16,8 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 from homeassistant.util import dt
 
-from . import HydrawiseEntity
 from .const import DATA_HYDRAWISE, LOGGER
+from .entity import HydrawiseEntity
 
 SENSOR_TYPES: tuple[SensorEntityDescription, ...] = (
     SensorEntityDescription(
